Capturing the fuzzy boundary of a plateau region object

This function yields a specific part of the fuzzy boundary of a plateau region object.

Usage

spa_boundary_pregion(pregion, bound_part = "region")

Arguments

pregion

A pgeometry object of the type PLATEAUREGION. It throws an error if a different type is given.

bound_part

A character value that indicates the part of the fuzzy boundary to be returned. It can be "region" or "line". See below for more details.

Value

A pgeometry object that represents a specific part of the fuzzy boundary of pgeometry object given as input.

Details

It employs the definition of fuzzy boundary of a fuzzy region object in the context of spatial plateau algebra (as defined in the references). The fuzzy boundary of a fuzzy region object A has a heterogeneous nature since it consists of two parts:

a fuzzy line object that corresponds to the boundary of the core of A.

a fuzzy region object that comprises all points of A with a membership degree greater than 0 and less than 1.

This means that the function spa_boundary_pregion can yield one specific part of the fuzzy boundary of a plateau region object (the argument pgeometry). If boundary = "line", then the function returns the boundary plateau line of pgeometry (i.e., returns a pgeometry object of the type PLATEAULINE). Else if boundary = "region" (the default value), then the function returns the boundary plateau region of pgeometry (i.e., returns a pgeometry object of the type PLATEAUREGION).

References

Carniel, A. C.; Schneider, M. A Conceptual Model of Fuzzy Topological Relationships for Fuzzy Regions. In Proceedings of the 2016 IEEE International Conference on Fuzzy Systems (FUZZ-IEEE 2016), pp. 2271-2278, 2016.

Carniel, A. C.; Schneider, M. Spatial Plateau Algebra: An Executable Type System for Fuzzy Spatial Data Types. In Proceedings of the 2018 IEEE International Conference on Fuzzy Systems (FUZZ-IEEE 2018), pp. 1-8, 2018.

Examples

library(tibble)
library(FuzzyR)

set.seed(123)

# some random points to create pgeometry objects by using the function spa_creator
tbl = tibble(x = runif(10, min= 0, max = 30), 
             y = runif(10, min = 0, max = 50), 
             z = runif(10, min = 0, max = 100))

classes <- c("category-1", "category-2")
mf1 <- genmf("trapmf", c(0, 5, 20, 35))
mf2 <- genmf("trimf", c(20, 80, 100))

pregions <- spa_creator(tbl, classes = classes, mfs = c(mf1, mf2))
pregions$pgeometry[[1]]
#> [1] "PLATEAUREGION ((POLYGON ((-10.8192 18.97073, 15.0174 25.2867, 25.22269 13.72177, 24.46465 12.93535, -10.8192 18.97073)), 0.405601751369735), (POLYGON ((12.21485 93.58036, 73.95271 93.58036, 73.95271 74.26182, 22.53832 39.81573, 11.0072 41.00587, 12.21485 93.58036)), 1))"
pregions$pgeometry[[2]]
#> [1] "PLATEAUREGION ((POLYGON ((73.95271 74.26182, 73.95271 20.30876, 41.31009 17.91326, 18.4713 28.79236, 22.53832 39.81573, 73.95271 74.26182)), 0.0286511168815196), (POLYGON ((-10.8192 18.97073, 15.0174 25.2867, 25.22269 13.72177, 24.46465 12.93535, -10.8192 18.97073)), 0.148599562157566), (POLYGON ((-13.51886 93.58036, 12.21485 93.58036, 11.0072 41.00587, 7.48209 40.08642, -13.51886 93.58036)), 0.552303419681266), (POLYGON ((-26.14517 -43.63571, -44.372 -43.63571, -44.372 5.185215, -14.69493 18.3287, -10.8192 18.97073, 24.46465 12.93535, 20.88759 6.753414, -26.14517 -43.63571)), 0.573443374518926), (POLYGON ((73.95271 -18.38102, 73.95271 -43.63571, -26.14517 -43.63571, 20.88759 6.753414, 73.95271 -18.38102)), 0.656903367411966), (POLYGON ((7.48209 40.08642, 11.0072 41.00587, 22.53832 39.81573, 18.4713 28.79236, 15.0174 25.2867, -10.8192 18.97073, -14.69493 18.3287, 7.48209 40.08642)), 0.734178022947162), (POLYGON ((73.95271 20.30876, 73.95271 -18.38102, 20.88759 6.753414, 24.46465 12.93535, 25.22269 13.72177, 41.31009 17.91326, 73.95271 20.30876)), 0.759509665193036), (POLYGON ((15.0174 25.2867, 18.4713 28.79236, 41.31009 17.91326, 25.22269 13.72177, 15.0174 25.2867)), 0.821339010260999), (POLYGON ((-44.372 5.185215, -44.372 93.58036, -13.51886 93.58036, 7.48209 40.08642, -14.69493 18.3287, -44.372 5.185215)), 0.847550780279562))"

# capturing and showing the boundary plateau line of each pgeometry object previously created
(spa_boundary_pregion(pregions$pgeometry[[1]], bound_part = "line")) 
#> [1] "PLATEAULINE ((LINESTRING (12.21485 93.58036, 73.95271 93.58036, 73.95271 74.26182, 22.53832 39.81573, 11.0072 41.00587, 12.21485 93.58036), 1))"
(spa_boundary_pregion(pregions$pgeometry[[2]], bound_part = "line"))
#> [1] "PLATEAULINE EMPTY"
# the last boundary is empty because there is no core! 

# capturing and showing the boundary plateau region (this is the default behavior)
(spa_boundary_pregion(pregions$pgeometry[[1]]))
#> [1] "PLATEAUREGION ((POLYGON ((-10.8192 18.97073, 15.0174 25.2867, 25.22269 13.72177, 24.46465 12.93535, -10.8192 18.97073)), 0.405601751369735))"
(spa_boundary_pregion(pregions$pgeometry[[2]]))
#> [1] "PLATEAUREGION ((POLYGON ((73.95271 74.26182, 73.95271 20.30876, 41.31009 17.91326, 18.4713 28.79236, 22.53832 39.81573, 73.95271 74.26182)), 0.0286511168815196), (POLYGON ((-10.8192 18.97073, 15.0174 25.2867, 25.22269 13.72177, 24.46465 12.93535, -10.8192 18.97073)), 0.148599562157566), (POLYGON ((-13.51886 93.58036, 12.21485 93.58036, 11.0072 41.00587, 7.48209 40.08642, -13.51886 93.58036)), 0.552303419681266), (POLYGON ((-26.14517 -43.63571, -44.372 -43.63571, -44.372 5.185215, -14.69493 18.3287, -10.8192 18.97073, 24.46465 12.93535, 20.88759 6.753414, -26.14517 -43.63571)), 0.573443374518926), (POLYGON ((73.95271 -18.38102, 73.95271 -43.63571, -26.14517 -43.63571, 20.88759 6.753414, 73.95271 -18.38102)), 0.656903367411966), (POLYGON ((7.48209 40.08642, 11.0072 41.00587, 22.53832 39.81573, 18.4713 28.79236, 15.0174 25.2867, -10.8192 18.97073, -14.69493 18.3287, 7.48209 40.08642)), 0.734178022947162), (POLYGON ((73.95271 20.30876, 73.95271 -18.38102, 20.88759 6.753414, 24.46465 12.93535, 25.22269 13.72177, 41.31009 17.91326, 73.95271 20.30876)), 0.759509665193036), (POLYGON ((15.0174 25.2867, 18.4713 28.79236, 41.31009 17.91326, 25.22269 13.72177, 15.0174 25.2867)), 0.821339010260999), (POLYGON ((-44.372 5.185215, -44.372 93.58036, -13.51886 93.58036, 7.48209 40.08642, -14.69493 18.3287, -44.372 5.185215)), 0.847550780279562))"