Visitation: An example of FSI model

visitation() provides an example, without rules, of a fuzzy spatial inference (FSI) model.

Usage

visitation()

Value

An FSI model without fuzzy rules set.

Details

The visitation() function provides a hypothetical FSI model that estimates the visiting experience based on prices and overall ratings of accommodations as well as sanitary conditions of restaurants. The output of such a model infers a value between 0 and 100 that indicates how attractive it is to visit a specific location. For this, the experience can be classified as awful, average, and great.

The linguistic variables and their linguistic values of this FSI model are listed below:

accommodation price with cut-rate, affordable, and expensive as linguistic values.

accommodation review with bad, good, and excellent as linguistic values.

food safety with low, medium, and high as linguistic values, which represent levels of sanitary conditions.

Note that this is just a small running example, containing a small set of points to represent the locations of accommodations and restaurants.

The usage of FSI models is subdivided into a preparation phase and an evaluation phase. The preparation phase is responsible for instantiating a new FSI model with the elements of the data source component of FIFUS. For this, the fsr package provides the following functions: fsi_create(), fsi_add_fsa(), and fsi_add_cs(). These functions are employed by visitation() so that users can add their own fuzzy set rules (by using fsi_add_rules()) and perform the evaluation phase (by using the functions fsi_eval() and/or fsi_qw_eval()).

In this sense, visitation() performs the following internal actions to return an FSI model:

specify the linguistic variables and their corresponding linguistic values, which are in turn represented by membership functions. These items are specified according to the context of the running example.

define small point datasets that represent each linguistic variable. Such datasets are tibble objects.

build spatial plateau objects by using spa_creator() on the datasets. As a result, we get spatial plateau objects that represent each linguistic value.

create an FSI model with fsi_create() function.

add fuzzy spatial antecedents with fsi_add_fsa(). Recall that the antecedents are spatial plateau objects previously built.

define the linguistic variable and its linguistic values with membership functions for the consequent.

add the consequent to the FSI model by using fsi_add_cs().

References

This function is based on the running example introduced in:

Carniel, A. C.; Galdino, F.; Philippsen, J. S.; Schneider, M. Handling Fuzzy Spatial Data in R Using the fsr Package. In Proceedings of the 29th International Conference on Advances in Geographic Information Systems (AM SIGSPATIAL 2021), pp. 526-535, 2021.

Underlying concepts and formal definitions of FIFUS are discussed in:

Carniel, A. C.; Schneider, M. Fuzzy inference on fuzzy spatial objects (FIFUS) for spatial decision support systems. In Proceedings of the 2017 IEEE International Conference on Fuzzy Systems (FUZZ-IEEE 2017), pp. 1-6, 2017.

Examples

fsi <- visitation()