Creation of an empty fuzzy spatial inference model

This function builds a fuzzy spatial inference (FSI) model without elements of the data source component (i.e., spatial plateau objects, fuzzy rules set, and fuzzy sets).

Usage

fsi_create(name, and_method = "min", or_method = "max",
           imp_method = "min", agg_method = "max", 
           defuzz_method = "centroid", default_conseq = NULL)

Arguments

name

A character value that specifies the name of the FSI model.

and_method

A character value that defines the operator name for the logical connective AND. Default value is "min".

or_method

A character value that defines the operator for the logical connective OR. Default value is "max".

imp_method

A character value that defines the operator for the implication operator. Default value is "min".

agg_method

A character value that defines the operator for the aggregation operator. Default value is "max".

defuzz_method

A character value that determines the defuzzification technique. Default value is the centroid technique.

default_conseq

This parameter is a membership function generated by the function genmf of the FuzzyR package.

Value

An empty named FSI model that is ready to be populated with fuzzy rules representing the antecedents and the consequent.

Details

The FSI model created with the function fsi_create and its default parameter values will implement a model using Mamdani's method. The possible values for the parameters and_method and imp_method are: "min", "prod". Other t-norms from the FuzzyR package are also conceivable. The possible value for the parameters or_method and agg_method is: "max". Other t-conorms from the FuzzyR package are also conceivable. The possible values for the parameter defuzz_method include other defuzzification techniques from the FuzzyR package. The parameter default_conseq defines the default behavior of the FSI model when there is no fuzzy rule with a degree of fulfillment greater than 0 returned by the FSI model.

After creating an empty FSI model, you have to call the functions fsi_add_fsa, fsi_add_cs, and fsi_add_rules to fulfill the FSI model.

Examples

library(FuzzyR)
# Creating the FSI model
fsi <- fsi_create("To visit or not to visit, that is the question", 
                  default_conseq = genmf("trimf", c(10, 30, 60)))